\name{Apr19}
\alias{Apr19}
\docType{data}
\title{April 2019}
\description{A data set containing the cleaned Divvy data for April 2019}
\usage{data("Apr19")}
\format{
  A data frame with 603 observations on the following 8 variables.
  \describe{
  \item{X}{the row number, automatically generated by exporting the CSV}
  \item{station_id}{the station ID number, from the Divvy program}
  \item{station_name}{the station name, from the Divvy program}
  \item{lat}{the station latitude}
  \item{long}{the station longitude}
  \item{total}{the total number of station rides in the given month}
  \item{electric}{the total number of station rides on E-bikes in the given month}
  \item{classic}{the total number of station rides on classic bikes in the given month}
  }
}
\details{Cleaned data for number of Divvy rides by station in April 2019,
subdivided by electric and classic bike, including all active network stations.}
\source{
\url{https://divvy-tripdata.s3.amazonaws.com/Divvy_Trips_2019_Q2.zip}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Apr19)
## maybe str(Apr19) ; plot(Apr19) ...
}
\keyword{datasets}
