\name{Jul15}
\alias{Jul15}
\docType{data}
\title{July 2015}
\description{A data set containing the cleaned Divvy data for July 2015}
\usage{data("Jul15")}
\format{
  A data frame with 474 observations on the following 8 variables.
  \describe{
  \item{X}{the row number, automatically generated by exporting the CSV}
  \item{station_id}{the station ID number, from the Divvy program}
  \item{station_name}{the station name, from the Divvy program}
  \item{lat}{the station latitude}
  \item{long}{the station longitude}
  \item{total}{the total number of station rides in the given month}
  \item{electric}{the total number of station rides on E-bikes in the given month}
  \item{classic}{the total number of station rides on classic bikes in the given month}
  }
}
\details{Cleaned data for number of Divvy rides by station in July 2015,
subdivided by electric and classic bike, including all active network stations.}
\source{
\url{https://divvy-tripdata.s3.amazonaws.com/Divvy_Trips_2015_Q3Q4.zip}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Jul15)
## maybe str(Jul15) ; plot(Jul15) ...
}
\keyword{datasets}
