\name{Sep14}
\alias{Sep14}
\docType{data}
\title{September 2014}
\description{A data set containing the cleaned Divvy data for September 2014}
\usage{data("Sep14")}
\format{
  A data frame with 300 observations on the following 8 variables.
  \describe{
  \item{X}{the row number, automatically generated by exporting the CSV}
  \item{station_id}{the station ID number, from the Divvy program}
  \item{station_name}{the station name, from the Divvy program}
  \item{lat}{the station latitude}
  \item{long}{the station longitude}
  \item{total}{the total number of station rides in the given month}
  \item{electric}{the total number of station rides on E-bikes in the given month}
  \item{classic}{the total number of station rides on classic bikes in the given month}
  }
}
\details{Cleaned data for number of Divvy rides by station in September 2014,
subdivided by electric and classic bike, including all active network stations.}
\source{
\url{https://divvy-tripdata.s3.amazonaws.com/Divvy_Stations_Trips_2014_Q3Q4.zip}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Sep14)
## maybe str(Sep14) ; plot(Sep14) ...
}
\keyword{datasets}
